add color change to health bar with DOTween
